pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "syncify"
version = "0.1.0"
description = "Spotify playlist synchronizer"
requires-python = ">=3.9"
dependencies = [
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "idna==3.11",
    "python-telegram-bot>=22.5",
    "pyyaml>=6.0.3",
    "requests==2.32.5",
    "types-requests==2.32.4.20260107",
    "urllib3==2.6.3",
]
